[fix] 시간 및 거리 int로 수정

+++ src/main/java/com/wayble/server/direction/dto/response/WayblePathResponse.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 @Builder
 public record WayblePathResponse(
-        double time,
-        double distance,
+        int time,
+        int distance,
         List<WayblePoint> points,
         List<double[]> polyline
 ) {
@@ -20,8 +20,8 @@ public record WayblePathResponse(
     ) {}
 
     public static WayblePathResponse of(
-            double time,
-            double distance,
+            int time,
+            int distance,
             List<WayblePoint> points,
             List<double[]> polyline
     ) {
